feat: 대시보드 업데이트 실패 시 재시도(#190)
